Register datastore/datapusher entry points
