+++ ckanext/datastore/cli.py
@@ -2,7 +2,6 @@
 
 import logging
 import os
-import re
 
 import click
